Add `Pool::train_population` and `Pool::get_best_genome`
The former can be used to provide "managed" training sessions for simple use cases where you can feed static training data and an evaluation function. This avoids the need to iterate over generations, species, genomes, etc. yourself.

Also included is a refactor of some of the `examples` code to move common things into a central place, and make use of the new built-in `TrainingData` structure.